deleteEnr refactor
