+++ core/src/main/java/samba/api/jsonrpc/done/PortalHistoryDeleteEnr.java
@@ -1,0 +1,41 @@
+package samba.api.jsonrpc.done;
+
+import samba.api.jsonrpc.parameters.InputsValidations;
+import samba.api.jsonrpc.parameters.ParametersUtil;
+import samba.api.libary.HistoryLibraryAPI;
+import samba.jsonrpc.config.RpcMethod;
+import samba.jsonrpc.reponse.JsonRpcMethod;
+import samba.jsonrpc.reponse.JsonRpcParameter;
+import samba.jsonrpc.reponse.JsonRpcRequestContext;
+import samba.jsonrpc.reponse.JsonRpcResponse;
+import samba.jsonrpc.reponse.JsonRpcSuccessResponse;
+import samba.network.history.api.HistoryNetworkInternalAPI;
+
+import java.util.Optional;
+
+public class PortalHistoryDeleteEnr implements JsonRpcMethod {
+
+    private final HistoryLibraryAPI historyLibraryAPI;
+
+    public PortalHistoryDeleteEnr(HistoryLibraryAPI historyLibraryAPI) {
+        this.historyLibraryAPI = historyLibraryAPI;
+    }
+
+    @Override
+    public String getName() {
+        return RpcMethod.PORTAL_HISTORY_DELETE_ENR.getMethodName();
+    }
+
+    @Override
+    public JsonRpcResponse response(JsonRpcRequestContext requestContext) {
+        try {
+            String nodeId = ParametersUtil.getNodeId(requestContext, 0);
+            boolean result = this.historyLibraryAPI.deleteEnr(nodeId);
+
+            return new JsonRpcSuccessResponse(requestContext.getRequest().getId(), result);
+
+        } catch (JsonRpcParameter.JsonRpcParameterException e) {
+            return createJsonRpcInvalidRequestResponse(requestContext);
+        }
+    }
+}
